feat: 헤더 별 페이지 분할
